move out the class to a new file
