+++ ProcessLib/ThermoHydroMechanics/ThermoHydroMechanicsFEM.h
@@ -24,6 +24,7 @@
 #include "ProcessLib/Parameter/Parameter.h"
 #include "ProcessLib/Utils/InitShapeMatrices.h"
 
+#include "IntegrationPointData.h"
 #include "ThermoHydroMechanicsProcessData.h"
 #include "LocalAssemblerInterface.h"
 
@@ -31,97 +32,6 @@ namespace ProcessLib
 {
 namespace ThermoHydroMechanics
 {
-template <typename BMatricesType, typename ShapeMatrixTypeDisplacement,
-          typename ShapeMatricesTypePressure, int DisplacementDim, int NPoints>
-struct IntegrationPointData final
-{
-    explicit IntegrationPointData(
-        MaterialLib::Solids::MechanicsBase<DisplacementDim> const&
-            solid_material)
-        : solid_material(solid_material),
-          material_state_variables(
-              solid_material.createMaterialStateVariables())
-    {
-    }
-
-    typename ShapeMatrixTypeDisplacement::template MatrixType<
-        DisplacementDim, NPoints * DisplacementDim>
-        N_u_op;
-    typename BMatricesType::KelvinVectorType sigma_eff, sigma_eff_prev;
-    typename BMatricesType::KelvinVectorType eps, eps_prev;
-    typename BMatricesType::KelvinVectorType eps_m, eps_m_prev;
-
-    typename ShapeMatrixTypeDisplacement::NodalRowVectorType N_u;
-    typename ShapeMatrixTypeDisplacement::GlobalDimNodalMatrixType dNdx_u;
-
-    typename ShapeMatricesTypePressure::NodalRowVectorType N_p;
-    typename ShapeMatricesTypePressure::GlobalDimNodalMatrixType dNdx_p;
-
-    MaterialLib::Solids::MechanicsBase<DisplacementDim> const& solid_material;
-    std::unique_ptr<typename MaterialLib::Solids::MechanicsBase<
-        DisplacementDim>::MaterialStateVariables>
-        material_state_variables;
-    double integration_weight;
-
-    void pushBackState()
-    {
-        eps_m_prev = eps_m;
-        sigma_eff_prev = sigma_eff;
-        material_state_variables->pushBackState();
-    }
-
-    template <typename DisplacementVectorType>
-    typename BMatricesType::KelvinMatrixType updateConstitutiveRelation(
-        double const t,
-        SpatialPosition const& x_position,
-        double const dt,
-        DisplacementVectorType const& /*u*/,
-        double const T)
-    {
-        auto&& solution = solid_material.integrateStress(
-            t, x_position, dt, eps_m_prev, eps_m, sigma_eff_prev,
-            *material_state_variables, T);
-
-        if (!solution)
-            OGS_FATAL("Computation of local constitutive relation failed.");
-
-        MathLib::KelvinVector::KelvinMatrixType<DisplacementDim> C;
-        std::tie(sigma_eff, material_state_variables, C) = std::move(*solution);
-
-        return C;
-    }
-
-    template <typename DisplacementVectorType>
-    typename BMatricesType::KelvinMatrixType updateConstitutiveRelationThermal(
-        double const t,
-        SpatialPosition const& x_position,
-        double const dt,
-        DisplacementVectorType const& /*u*/,
-        double const T,
-        double const thermal_strain)
-    {
-        auto const& identity2 = MathLib::KelvinVector::Invariants<
-            MathLib::KelvinVector::KelvinVectorDimensions<
-                DisplacementDim>::value>::identity2;
-
-        // assume isotropic thermal expansion
-        eps_m.noalias() = eps - thermal_strain * identity2;
-        auto&& solution = solid_material.integrateStress(
-            t, x_position, dt, eps_m_prev, eps_m, sigma_eff_prev,
-            *material_state_variables, T);
-
-        if (!solution)
-            OGS_FATAL("Computation of local constitutive relation failed.");
-
-        MathLib::KelvinVector::KelvinMatrixType<DisplacementDim> C;
-        std::tie(sigma_eff, material_state_variables, C) = std::move(*solution);
-
-        return C;
-    }
-
-    EIGEN_MAKE_ALIGNED_OPERATOR_NEW;
-};
-
 /// Used by for extrapolation of the integration point values. It is ordered
 /// (and stored) by integration points.
 template <typename ShapeMatrixType>
