+++ Tests/Data/ThermoHydroMechanics/Linear/Beam_sealed_bimaterial/expected_square_1e2_pcs_0_ts_10_t_100.000000.vtu
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:7d5af5b0ea577a4091a38a32db128be0811f5dea25494d5a2525f4c7dbe3677b
-size 351565
+oid sha256:ae640efd0d16acc0cfd7175e9d692690fc9b151fecb869a21a83881f2c3842c4
+size 356872
