+++ Tests/Data/ThermoHydroMechanics/Linear/Beam_unsealed_bimaterial/expected_square_1e2_pcs_0_ts_10_t_1000.000000.vtu
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f0f2f4db5fd744d0a1046f7fa56f621cfc08d0069169aa90242964bff7792b34
-size 329569
+oid sha256:0e79fa4019cf033b3f93f65cb56b37a0666e7a89ceaf510b9e9f17159c28a82a
+size 335560
