+++ Tests/Data/ThermoHydroMechanics/Linear/Point_injection/expected_square_1e0_pcs_0_ts_10_t_10000.000000.vtu
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:209160d37cea086e0f963a6f00785d0fb7a7e60de8e2119de068eacefcdea166
-size 436969
+oid sha256:4db85b7114211706406e3965d24be6b82c6ca140f40b1dd6cf606aaf8aa00036
+size 460041
