--- Tests/Data/ThermoHydroMechanics/Linear/Square_sealed_homogeneous/analytical.vtu
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:520787ad847fbb2295c78f5265efd1764d0e76dea41c979b2be18d08c9208c56
-size 5043
