+++ Tests/Data/ThermoHydroMechanics/Linear/Square_sealed_homogeneous/expected_square_1e0_pcs_0_ts_10_t_1000.000000.vtu
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:dc9dcd6793620b74872ec3c6380f5ebf050dc0ccae90ba0ad90345be4b96a062
-size 3884
+oid sha256:8b4b9b1241971805a4aee21f54ef55429e22e1e526ba453930571621536ad55a
+size 3786
