+++ web/content/docs/benchmarks/thermo-hydro-mechanics/THM.pdf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:4b794d2e505398b8933a6fa7064af9a2109b2eec45a2bda14249ea8dcaf9fcd2
-size 503230
+oid sha256:8cd4be7769ca207a3abe8f01015bbd7765f7c902969ee38972b5334800c49c72
+size 502372
